signal_server/requirements.txt
python-socketio
uvicorn[standard]